prodcon/requirements.txt
# black==21.12b0
# click==8.0.3
kafka-python
faker
# mypy-extensions==0.4.3
# pathspec==0.9.0
# platformdirs==2.4.0
# tomli==1.2.2
# typing-extensions==4.0.1
